bump

+++ SecretAdmin/Program.cs
@@ -1,12 +1,11 @@
 ﻿using System;
-using System.Drawing;
 using System.IO;
-using System.Threading.Tasks;
+using SecretAdmin.API;
 using SecretAdmin.Features.Console;
 using SecretAdmin.Features.Program;
+using SecretAdmin.Features.Program.Config;
 using SecretAdmin.Features.Server;
 using SecretAdmin.Features.Server.Commands;
-using SecretAdmin.Features.Server.Enums;
 
 namespace SecretAdmin
 {
@@ -14,27 +13,38 @@ namespace SecretAdmin
     {
         public static Version Version { get; } = new (0, 0, 0,1);
         public static ScpServer Server { get; private set; }
+        public static ConfigManager ConfigManager { get; private set; }
+        public static Logger ProgramLogger { get; private set; }
         public static CommandHandler CommandHandler { get; private set; }
 
         static void Main(string[] args)
         {
             AppDomain.CurrentDomain.ProcessExit += OnExit;
-            
             Console.Title = $"SecretAdmin [v{Version}]";
 
-            Log.Intro();
-            Console.ReadKey();
+            var arguments = ArgumentsManager.GetArgs(args);
             
-            if (ProgramIntroduction.FirstTime)
-                ProgramIntroduction.ShowIntroduction();
-
             Paths.Load();
+            ConfigManager = new ConfigManager();
+            if (ProgramIntroduction.FirstTime || arguments.Reconfigure)
+                ProgramIntroduction.ShowIntroduction();
+            ConfigManager.LoadConfig();
             
-            CommandHandler = new CommandHandler();
-            
-            Server = new ScpServer(7777);
-            Server.Start();
+            if(arguments.Logs)
+                ProgramLogger = new Logger(Path.Combine(Paths.ProgramLogsFolder, $"{DateTime.Now:MM.dd.yyyy-hh.mm.ss}.log"));
 
+            if(ConfigManager.SecretAdminConfig.AutoUpdater)
+                AutoUpdater.CheckForUpdates();
+            
+            Log.Intro();
+            Utils.ArchiveControlLogs();
+
+            CommandHandler = new CommandHandler();
+            ModuleManager.LoadAll();
+            
+            Server = new ScpServer(ConfigManager.GetServerConfig(arguments.Config));
+            Server.Start();
+            
             InputManager.Start();
         }
 
@@ -43,8 +53,12 @@ namespace SecretAdmin
             Console.ForegroundColor = ConsoleColor.Blue;
             Console.WriteLine("\nExit Detected. Killing game process.");
 
-            Server?.Kill();
-            
+            if (ConfigManager.SecretAdminConfig.SafeShutdown)
+                Server?.Kill();
+
+            foreach (var module in ModuleManager.Modules)
+                module.OnDisabled();
+
             Console.ForegroundColor = ConsoleColor.Cyan;
             Console.WriteLine("Everything seems good to go! Bye :)");
         }
